% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pBlackNetworkAnalysis.R
\docType{data}
\name{allKGML.RData}
\alias{allKGML.RData}
\title{allKGML.RData}
\source{
downloaded using KEGGREST from KEGG on June 4, 2014
}
\description{
allKGML.RData
}
\details{
contains set of KGML xml data that P Blacks genes were annotated to as of June 4, 2014
}

