% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R, R/pBlackNetworkAnalysis.R
\docType{class}
\name{annotation-class}
\alias{annotation-class}
\title{annotation class}
\description{
annotation class

annotation class
}
\section{Slots}{

\describe{
\item{\code{annotation_to_feature}}{list where each named entry is the features annotated to it}

\item{\code{description}}{named character vector}

\item{\code{link}}{named character vector}

\item{\code{annotation_to_feature}}{list where each named entry is the features annotated to it}

\item{\code{description}}{named character vector}

\item{\code{link}}{named character vector}
}}

