% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R
\name{black.miRNAmRNA.NSCLC.networkEnrichment}
\alias{black.miRNAmRNA.NSCLC.networkEnrichment}
\title{package black.miRNAmRNA.NSCLC.networkEnrichment}
\description{
Package used to generate STRING based protein-protein interactions and perform
network enrichment using KEGG pathways and GO biological process for the publication:
W Wu, RM Flight, MJ Krentz, B Kulengowski, H-F Li, HNB Moseley, A Stromberg and EP Black,
\emph{Interacting miRNA and mRNA genes identify potential therapeutic targets for NSCLC},
sumbitted Feb 2015.
}

