% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R
\name{collapseNodes}
\alias{collapseNodes}
\title{collapse nodes and info}
\usage{
collapseNodes(ppiGraph, ppiInfo, ppiSeed, collapseString)
}
\arguments{
\item{ppiGraph}{the graph of interactions}

\item{ppiInfo}{the information about each node}

\item{ppiSeed}{the nodes that were originally used to seed the PPI graph}

\item{collapseString}{the strings used to find those things to collapse}
}
\value{
list
}
\description{
sometimes our graph has a lot of rather duplicated nodes that we want to collapse for visualization.
thats what this does. Given a graph, the info about the nodes, and the strings to use to find things to
collapse together, returns a new graph and info entries.
}

