% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R, R/pBlackNetworkAnalysis.R
\name{communityAsList}
\alias{communityAsList}
\title{spit out community structure}
\usage{
communityAsList(commObject, decreasing = TRUE)

communityAsList(commObject, decreasing = TRUE)
}
\arguments{
\item{commObject}{the community object returned by an igraph community calculation}

\item{decreasing}{sort the communities with largest first (TRUE, default) or last (FALSE)}

\item{commObject}{the community object returned by an igraph community calculation}

\item{decreasing}{sort the communities with largest first (TRUE, default) or last (FALSE)}
}
\value{
list of communities

list of communities
}
\description{
spit out community structure

spit out community structure
}

