% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R
\name{getPPI}
\alias{getPPI}
\title{get PPI network from STRING}
\usage{
getPPI(seedGenes, stringDB)
}
\arguments{
\item{seedGenes}{the genes to seed the PPI}

\item{stringDB}{the STRING data.frame to use}
}
\value{
data.frame
}
\description{
Given a set of gene identifiers and a STRING data.frame, get the interactors within
a single edge of the \emph{seed} genes, and the set of all edges between the genes
}

