% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R, R/pBlackNetworkAnalysis.R
\docType{class}
\name{hypergeom_features-class}
\alias{hypergeom_features-class}
\title{hypergeom feature class}
\description{
class to hold features undergoing hypergeometric enrichment

class to hold features undergoing hypergeometric enrichment
}
\section{Slots}{

\describe{
\item{\code{significant}}{the significant features}

\item{\code{universe}}{all of the features measured}

\item{\code{annotation}}{annotation object}

\item{\code{significant}}{the significant features}

\item{\code{universe}}{all of the features measured}

\item{\code{annotation}}{annotation object}
}}

