% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R, R/pBlackNetworkAnalysis.R
\name{hypergeometric_feature}
\alias{hypergeometric_feature}
\title{do hypergeometric enrichment}
\usage{
hypergeometric_feature(hypergeom_features, direction = "over")

hypergeometric_feature(hypergeom_features, direction = "over")
}
\arguments{
\item{hypergeom_features}{a hypergeometric_features object}

\item{direction}{which direction to do the enrichment (over or under)}

\item{hypergeom_features}{a hypergeometric_features object}

\item{direction}{which direction to do the enrichment (over or under)}
}
\value{
hypergeometric_Feature

hypergeometric_Feature
}
\description{
do hypergeometric enrichment

do hypergeometric enrichment
}

