% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R, R/pBlackNetworkAnalysis.R
\docType{data}
\name{kegg_annotation.RData}
\alias{kegg_annotation.RData}
\title{kegg_annotation.RData}
\source{
downloaded using KEGGREST on Feb 11, 2015

downloaded using KEGGREST on Feb 11, 2015
}
\description{
kegg_annotation.RData

kegg_annotation.RData
}
\details{
list with annotation contains the gene - pathway information, and pathway descriptions

list with annotation contains the gene - pathway information, and pathway descriptions
}
\examples{
\dontrun{
library(KEGGREST)
hsa_pathways <- keggLink("hsa", "pathway")
names(hsa_pathways) <- substring(names(hsa_pathways), 6)
hsa_pathways <- substring(hsa_pathways, 5)
kegg_annotation <- list(annotation = split(hsa_pathways, names(hsa_pathways)))
kegg_annotation$annotation <- lapply(kegg_annotation$annotation, function(x){names(x) <- NULL; x})

kegg_paths <- names(kegg_annotation$annotation)
kegg_desc <- keggList("pathway", "hsa")
names(kegg_desc) <- substring(names(kegg_desc), 6)
kegg_annotation$description <- kegg_desc[names(kegg_annotation$annotation)]
}
\dontrun{
library(KEGGREST)
hsa_pathways <- keggLink("hsa", "pathway")
names(hsa_pathways) <- substring(names(hsa_pathways), 6)
hsa_pathways <- substring(hsa_pathways, 5)
kegg_annotation <- list(annotation = split(hsa_pathways, names(hsa_pathways)))
kegg_annotation$annotation <- lapply(kegg_annotation$annotation, function(x){names(x) <- NULL; x})

kegg_paths <- names(kegg_annotation$annotation)
kegg_desc <- keggList("pathway", "hsa")
names(kegg_desc) <- substring(names(kegg_desc), 6)
kegg_annotation$description <- kegg_desc[names(kegg_annotation$annotation)]
}
}

