% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R, R/pBlackNetworkAnalysis.R
\name{map2STRING}
\alias{map2STRING}
\title{map to STRING}
\usage{
map2STRING(queryData, queryCol, stringAliasFile, removeUnmapped = TRUE)

map2STRING(queryData, queryCol, stringAliasFile, removeUnmapped = TRUE)
}
\arguments{
\item{stringAliasFile}{the file with STRING aliases}

\item{query}{what you want to match on}

\item{query}{what you want to match on}

\item{stringAliasFile}{the file with STRING aliases}
}
\value{
data.frame of translated queries

data.frame of translated queries
}
\description{
map to STRING

map to STRING
}

