% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R
\docType{data}
\name{PPI_seed.RData}
\alias{PPI_seed.RData}
\title{PPI_seed.RData}
\source{
filtered STRING PPI network with edges > 400, seeded from EIF4B, RAB8A, and PPPC1B
}
\description{
PPI_seed.RData
}

