% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R
\name{PPI_to_graph}
\alias{PPI_to_graph}
\title{graph from PPI}
\usage{
PPI_to_graph(PPI, orgDB)
}
\arguments{
\item{PPI}{the STRING interaction network}

\item{orgDB}{an organism database to add data to the network}
}
\value{
list
}
\description{
generates a graph from the PPI network suitable for visualization
}

