% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirna_mrna_network_enrichment.R
\docType{data}
\name{seedMRNA.RData}
\alias{seedMRNA.RData}
\title{seedMRNA.RData}
\source{
the \code{outMRNA_mapped} object for the seed genes of EIF4B, RAB8A, and PPPC1B
}
\description{
seedMRNA.RData
}

